requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
confuse==2.0.1
dnspython==2.7.0
idna==3.10
numpy==2.0.2
pandas==2.2.3
pymongo==4.10.1
python-amazon-sp-api==1.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0